requirements.txt
fastapi==0.63.0  # version fixed due to complete re-overriding fastapi objects
pydantic  # version fixed by fastapi
python-dotenv
uvicorn
sqlalchemy
psycopg2-binary
